// src/hooks/types.ts
export const statisticOptions = [
  { value: 'sum', label: 'Sum' },
  { value: 'mean', label: 'Mean' },
  { value: 'std', label: 'Standard Deviation' },
  { value: 'min', label: 'Minimum' },
  { value: 'max', label: 'Maximum' },
  { value: 'median', label: 'Median' },
  { value: 'count', label: 'Count' },
  { value: 'range', label: 'Range' },
  { value: 'available_data', label: 'Available Data' },
] as const

export type StatisticType = (typeof statisticOptions)[number]['value']
// ...
export interface KPIInstanceProps {
  name_short: string
  name_long: string
  unit: string
  kpi_type_id: number
  description?: string
  aggregation_method: string
  device_type_id: number
}

export interface KPIData {
  x: string[]
  y: Array<{
    device_values: { [deviceId: string]: number | null }
  }>
}

export interface alertProps {
  alert_name: string
  comparison: string | null
  duration_value: string | null
  kpi_type_id: string | null
  statistic: StatisticType | null
  notify: boolean
  threshold_value: number | null | string
  triggered: boolean | null
}

export interface KPIAlertProps {
  kpi_alert_id: number
  user_id: string
  project_id: string
  kpi_type_id: string
  config: alertProps
}